add tests 93% + black + mypy + isort
